Added badges feature

+++ backend/scripts/editReview.py
@@ -11,6 +11,7 @@ import json
 
 from scripts.adminFunctions import hash_password
 from scripts.createReview import create_username
+from scripts import badge_helpers
 
 file_name = os.path.basename(__file__)
 blueprint = Blueprint(file_name[:-3], __name__)
@@ -131,15 +132,14 @@ def updateReview(id):
     cur.execute("""SELECT * FROM "pointSystemRules" WHERE "id" BETWEEN 2 AND 6""")
     point_system_rules = cur.fetchall()
 
-
     # Check the difference between the new review and the existing review
     remove_component = []
     added_component = []
 
     # [1] Check if text review was removed
-    if (data['reviewDesc'] == '' and data['reviewDesc'] is None) and (existing_review['reviewDesc'] != '' and existing_review['reviewDesc'] is not None):
+    if (data['reviewDesc'] == '' or data['reviewDesc'] is None) and (existing_review['reviewDesc'] != '' and existing_review['reviewDesc'] is not None):
         remove_component.append(2)
-    elif (data['reviewDesc'] != '') and (existing_review['reviewDesc'] == '' and existing_review['reviewDesc'] is None):
+    elif (data['reviewDesc'] != '' and data['reviewDesc'] is not None) and (existing_review['reviewDesc'] == '' or existing_review['reviewDesc'] is None):
         added_component.append(2)
 
     # [2] Check if extensive review was removed or added
@@ -153,7 +153,7 @@ def updateReview(id):
     current_review_ext_taste = bool(existing_review.get('taste'))
     current_review_ext_finish = bool(existing_review.get('finish'))
 
-   # Count how many fields exist in current and updated review
+    # Count how many fields exist in current and updated review
     current_count = sum([
         current_review_ext_color,
         current_review_ext_aroma,
@@ -168,30 +168,41 @@ def updateReview(id):
         updated_review_ext_finish
     ])
 
+    # Track if extensive review was added or removed
+    is_extensive_review_before = current_count > 0
+    is_extensive_review_after = updated_count > 0
+
     # Check if anything was removed or added
-    if updated_count > 0 and current_count == 0:
+    if is_extensive_review_after and not is_extensive_review_before:
         added_component.append(3)
-    elif updated_count == 0 and current_count > 0:
+    elif not is_extensive_review_after and is_extensive_review_before:
         remove_component.append(3)
     
-
-    # [3] Check if photo was removed or added - not working
-    if is_empty_photo(data['photo']) and is_non_empty_photo(existing_review['photo']):
+    # [3] Check if photo was removed or added
+    has_photo_before = is_non_empty_photo(existing_review['photo'])
+    has_photo_after = is_non_empty_photo(data['photo'])
+    
+    if not has_photo_after and has_photo_before:
         remove_component.append(4)
-    # Check if photo was added
-    elif is_non_empty_photo(data['photo']) and is_empty_photo(existing_review['photo']):
+    elif has_photo_after and not has_photo_before:
         added_component.append(4)
 
     # [4] Check if location was removed or added
-    if (data['location'] == '' and data['location'] is None) and (existing_review['location'] != '' and existing_review['location'] is not None):
+    has_location_before = existing_review['location'] is not None and existing_review['location'] != ''
+    has_location_after = data['location'] is not None and data['location'] != ''
+    
+    if not has_location_after and has_location_before:
         remove_component.append(5)
-    elif (data['location'] != '') and (existing_review['location'] == '' and existing_review['location'] is None):
+    elif has_location_after and not has_location_before:
         added_component.append(5)
 
     # [5] Check if tagged users were removed or added
-    if (data['taggedUsers'] == []) and (existing_review['taggedUsers'] != []):
+    has_tagged_friends_before = existing_review['taggedUsers'] != []
+    has_tagged_friends_after = data['taggedUsers'] != []
+    
+    if not has_tagged_friends_after and has_tagged_friends_before:
         remove_component.append(6)
-    elif (data['taggedUsers'] != []) and (existing_review['taggedUsers'] == []):
+    elif has_tagged_friends_after and not has_tagged_friends_before:
         added_component.append(6)
 
     # Insert or find the venue
@@ -217,9 +228,9 @@ def updateReview(id):
             conn.commit()
 
     # Update review photo
-    if existing_review['photo']:
+    if existing_review['photo'] and data['photo'] != existing_review['photo']:
         s3Images.deleteImageFromS3(existing_review['photo'])
-    if data['photo']:
+    if data['photo'] and data['photo'] != existing_review['photo']:
         data['photo'] = s3Images.uploadBase64ImageToS3(data['photo'])
 
     tagged_users = data.get('taggedUsers', [])
@@ -263,15 +274,78 @@ def updateReview(id):
             conn.commit()
 
             print("Additional points: ", modify_point)
-
+            
+        user_id = data.get('userID')
+        badges_updated = []
+        
+        # Fetch listing info to get drink type, category and country data
+        cur.execute("""
+            SELECT "drinkType", "typeCategory", "originCountry" 
+            FROM "listings" 
+            WHERE id = %s
+        """, (data.get('reviewTarget'),))
+        
+        listing_info = cur.fetchone()
+        
+        if listing_info and (added_component or remove_component):
+            
+            if 3 in added_component:
+                # ExtensiveReview added - increase badge progress
+                badge_updated = badge_helpers.update_badge_progress(conn, cur, user_id, 'ExtensiveReview', 'Action', 1)
+                if badge_updated:
+                    badges_updated.append(badge_updated)
+            elif 3 in remove_component:
+                # ExtensiveReview removed - decrease badge progress
+                badge_updated = badge_helpers.update_badge_progress(conn, cur, user_id, 'ExtensiveReview', 'Action', -1)
+                if badge_updated:
+                    badges_updated.append(badge_updated)
+                    
+            # 2. Process PhotoAttached badge
+            if 4 in added_component:
+                # Photo added - increase badge progress
+                badge_updated = badge_helpers.update_badge_progress(conn, cur, user_id, 'PhotoAttached', 'Action', 1)
+                if badge_updated:
+                    badges_updated.append(badge_updated)
+            elif 4 in remove_component:
+                # Photo removed - decrease badge progress
+                badge_updated = badge_helpers.update_badge_progress(conn, cur, user_id, 'PhotoAttached', 'Action', -1)
+                if badge_updated:
+                    badges_updated.append(badge_updated)
+                    
+            # 3. Process LocationTagged badge
+            if 5 in added_component:
+                # Location added - increase badge progress
+                badge_updated = badge_helpers.update_badge_progress(conn, cur, user_id, 'LocationTagged', 'Action', 1)
+                if badge_updated:
+                    badges_updated.append(badge_updated)
+            elif 5 in remove_component:
+                # Location removed - decrease badge progress
+                badge_updated = badge_helpers.update_badge_progress(conn, cur, user_id, 'LocationTagged', 'Action', -1)
+                if badge_updated:
+                    badges_updated.append(badge_updated)
+                    
+            # 4. Process FriendTagged badge
+            if 6 in added_component:
+                # Friends tagged - increase badge progress
+                badge_updated = badge_helpers.update_badge_progress(conn, cur, user_id, 'FriendTagged', 'Action', 1)
+                if badge_updated:
+                    badges_updated.append(badge_updated)
+            elif 6 in remove_component:
+                # Friends untagged - decrease badge progress
+                badge_updated = badge_helpers.update_badge_progress(conn, cur, user_id, 'FriendTagged', 'Action', -1)
+                if badge_updated:
+                    badges_updated.append(badge_updated)
+        
         return jsonify({
             "code": 200,
             "data": data.get('reviewDesc', ''),
             "pointsChange": modify_point,
+            "badgesUpdated": badges_updated
         }), 200
 
     except Exception as e:
         print(str(e))
+        conn.rollback()
         return jsonify({
             "code": 500,
             "data": {
